requirements.txt
Flask==3.0.3
google-genai==1.28.0
flask-cors==6.0.1